Use graphql on compiler, executor and verdict
Add `submission` graphql query.
Now compiler, executor and verdict only receive a submission ID and get
all information they need from graphql.
Using github.com/hasura/go-graphql-client to handle graphql queries.
Create volume shared by compiler and executor to store binary file from
a submission.
TimeLimit is now a float32 instead of an integer and is actually used by
executor.
This breaks the connection between orchestrator and verdict since
orchestrator does not send a submission ID to verdict.

+++ model/model.go
@@ -29,7 +29,7 @@ type Challenge struct {
 	gorm.Model
 	ID          uint
 	Title       string
-	TimeLimit   int
+	TimeLimit   float32
 	MemoryLimit int
 	Inputs      []TestFile `gorm:"ForeignKey:ChallengeID"`
 	Outputs     []TestFile `gorm:"ForeignKey:ChallengeID"`
